backend/requirements.txt
asgiref==3.8.1
Django>=4.2
sqlparse==0.5.3
typing_extensions==4.13.2
mysqlclient>=2.1
celery==5.3.6
djangorestframework==3.15.0
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.2.2


